pyproject.toml
[project]
name = "codingassist-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.0",
    "openai==1.54.0",
]